# api/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "SamaCloud",
        "version": "0.1.0",
        "description": "API de contr\u00f4le de SamaCloud : comptes, jetons, projets, d\u00e9ploiements, commandes et \u00e9change.\n\nCette sp\u00e9cification est une **proposition d'API pour Datacloud**, \u00e9crite pour le concours CADEV. Ce n'est pas l'API officielle de Systalink.\n\nToutes les r\u00e9ponses de succ\u00e8s sont envelopp\u00e9es dans `data`. Toutes les erreurs ont la forme `{ code, message, details? }`, avec un message en fran\u00e7ais."
    },
    "servers": [
        {
            "url": "/api",
            "description": "Serveur qui h\u00e9berge cette documentation"
        }
    ],
...
    "security": [
        {
            "jeton": []
        }
    ],
    "paths": {
...
        "/questions/{id}/signalements": {
            "post": {
                "operationId": "reports.question",
                "description": "Pr\u00e9vient les administrateurs qu'une question pose probl\u00e8me. Signaler deux fois\nla m\u00eame question ne cr\u00e9e pas de doublon.",
                "summary": "Signaler une question",
                "tags": [
                    "Communaut\u00e9 \u2014 questions-r\u00e9ponses"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "Identifiant de la question.",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ReportRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "`ReportResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/ReportResource"
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "- `non_authentifie` \u2014 Jeton absent, expir\u00e9 ou r\u00e9voqu\u00e9.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Erreur"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "- `compte_suspendu` \u2014 Le compte a \u00e9t\u00e9 suspendu par un administrateur.\n- `capacite_manquante` \u2014 Le jeton n'a pas la capacit\u00e9 exig\u00e9e par cette route.\n- `signalement_de_son_contenu` \u2014 On ne signale pas son propre contenu.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Erreur"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "- `introuvable` \u2014 La ressource n'existe pas ou n'appartient pas \u00e0 ce compte.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Erreur"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "- `validation_echouee` \u2014 Donn\u00e9es invalides ; \u00ab details \u00bb liste chaque champ fautif.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Erreur"
                                }
                            }
                        }
                    },
                    "429": {
                        "description": "- `trop_de_requetes` \u2014 Limite de d\u00e9bit atteinte ; voir \u00ab details.reessayer_dans_secondes \u00bb.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Erreur"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "jeton": {
                "type": "http",
                "description": "Jeton de session (`sc_sess_\u2026`) ou jeton d'agent IA (`sc_live_\u2026`, `sc_test_\u2026`).",
                "scheme": "bearer"
            }
        },
        "schemas": {
...
            "AuthorResource": {
                "type": "object",
                "properties": {
                    "pseudo": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "description": "\u00c0 utiliser avec GET /api/profils/{pseudo}."
                    },
                    "nom": {
                        "type": "string"
                    },
                    "pays": {
                        "type": [
                            "string",
                            "null"
                        ]
                    }
                },
                "required": [
                    "pseudo",
                    "nom",
                    "pays"
                ],
                "title": "AuthorResource"
            },
...
            "Erreur": {
                "type": "object",
                "properties": {
                    "code": {
                        "type": "string",
                        "description": "Identifiant stable de l'erreur, \u00e0 lire par les programmes.",
                        "examples": [
                            "quota_atteint"
                        ]
                    },
                    "message": {
                        "type": "string",
                        "description": "Explication en fran\u00e7ais, affichable telle quelle."
                    },
                    "details": {
                        "description": "Pr\u00e9cisions \u00e9ventuelles. Pour une erreur de validation : liste de { champ, message }."
                    }
                },
                "required": [
                    "code",
                    "message"
                ],
                "title": "Erreur"
            },
...
            "ReportReason": {
                "type": "string",
                "description": "Motif d'un signalement.\n",
                "enum": [
                    "spam",
                    "contenu_offensant",
                    "hors_sujet",
                    "autre"
                ],
                "title": "ReportReason"
            },
            "ReportRequest": {
                "type": "object",
                "description": "Signalement d'un contenu.",
                "properties": {
                    "motif": {
                        "$ref": "#/components/schemas/ReportReason"
                    },
                    "details": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "description": "Pr\u00e9cisions facultatives pour l'administrateur ; obligatoires si le motif est \u00ab autre \u00bb.",
                        "maxLength": 500
                    }
                },
                "required": [
                    "motif"
                ],
                "title": "ReportRequest"
            },
            "ReportResource": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "type": {
                        "description": "\u00ab question \u00bb ou \u00ab reponse \u00bb.",
                        "$ref": "#/components/schemas/ReportableType"
                    },
                    "contenu_id": {
                        "type": "string"
                    },
                    "motif": {
                        "$ref": "#/components/schemas/ReportReason"
                    },
                    "details": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "statut": {
                        "$ref": "#/components/schemas/ReportStatus"
                    },
                    "signale_par": {
                        "$ref": "#/components/schemas/AuthorResource"
                    },
                    "cree_le": {
                        "type": "string"
                    },
                    "traite_le": {
                        "type": [
                            "string",
                            "null"
                        ]
                    }
                },
                "required": [
                    "id",
                    "type",
                    "contenu_id",
                    "motif",
                    "details",
                    "statut",
                    "cree_le",
                    "traite_le"
                ],
                "title": "ReportResource"
            },
            "ReportStatus": {
                "type": "string",
                "description": "\u00c9tat d'un signalement.\n| |\n|---|\n| `ouvert` <br/> En attente de la d\u00e9cision d'un administrateur. |\n| `retenu` <br/> Signalement justifi\u00e9 : le contenu a \u00e9t\u00e9 retir\u00e9. |\n| `rejete` <br/> Signalement non justifi\u00e9 : le contenu reste en ligne. |",
                "enum": [
                    "ouvert",
                    "retenu",
                    "rejete"
                ],
                "title": "ReportStatus"
            },
            "ReportableType": {
                "type": "string",
                "description": "Types de contenus qu'un membre peut signaler.\n",
                "enum": [
                    "question",
                    "reponse"
                ],
                "title": "ReportableType"
            },
...
        }
    }
}